music connect

+++ 쇼핑몰 기능 추가 테스트본/src/main/java/edu/du/sb1024/mapper/OrderMapper.java
@@ -27,7 +27,7 @@ public interface OrderMapper {
 
     void updateOrder(OrderDto order) throws Exception;
 
-    void deleteOrder(int orderIdx) throws Exception;
+    void deleteOrder(Long orderIdx) throws Exception;
 
     void insertOrderFileList(List<OrderFileDto> list) throws Exception;
